+++ authlete/dto/client.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2019 Authlete, Inc.
+# Copyright (C) 2019-2020 Authlete, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -56,7 +56,8 @@ class Client(Jsonable):
             'tosUris':                               TaggedValue,      # list of TaggedValue
             'jwksUri':                               str,
             'jwks':                                  str,
-            'sectorIdentifier':                      str,
+            'derivedSectorIdentifier':               str,
+            'sectorIdentifierUri':                   str,
             'subjectType':                           SubjectType,
             'idTokenSignAlg':                        JWSAlg,
             'idTokenEncryptionAlg':                  JWEAlg,
@@ -96,7 +97,10 @@ class Client(Jsonable):
             'bcRequestSignAlg':                      JWSAlg,
             'bcUserCodeRequired':                    bool,
             'dynamicallyRegistered':                 bool,
-            'registrationAccessTokenHash':           str
+            'registrationAccessTokenHash':           str,
+            'authorizationDataTypes':                str,              # list of str
+            'parRequired':                           bool,
+            'requestObjectRequired':                 bool,
         }
 
         super().__init__(nameAndValues, nameAndTypes)